KafkaThesisBuild/kafka/Clients/BackEndService/requirements.txt
asyncio
confluent-kafka
websockets
python-dotenv
requests
fastavro
